lessons/02-usecases/code/server/requirements.txt
# requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
pydantic
python-dotenv
